Updated sslswap()
This corrects for an empty .chain.crt, if this is empty it runs the
Interworx command to install only the .crt and not the chain.crt.

+++ bashrc.sh
@@ -1406,7 +1406,7 @@ if [[ $keyhash != $crthash ]]; then
 
 else
   echo -e "\n[${BRIGHT}${GREEN}UPDATE${NORMAL}] .. SSL Certificate"
-  rm ${domain}.crt; mv ${domain}{.new.crt,.crt}
+  rm ${domain}.crt 2> /dev/null; mv ${domain}{.new.crt,.crt}
   chmod 600 ${domain}.crt; chown iworx. ${domain}.crt
 
   # Check if new chain cert exists and is non-zero; then remove and replace the old one
@@ -1416,10 +1416,15 @@ else
     chmod 600 ${domain}.chain.crt; chown iworx. ${domain}.chain.crt
   fi
 
-  sudo -u $(getusr) siteworx -unc Ssl -a install --domain $domain --chain 1
+  # Check if new chain cert exists and is non-zero; then install new SSL with Chain, else exclude chain
+  if [[ -f ${domain}.new.chain.crt && -n $(cat ${domain}.new.chain.crt 2> /dev/null) ]]; then
+    sudo -u $(getusr) siteworx -unc Ssl -a install --domain $domain --chain 1
+  else
+    sudo -u $(getusr) siteworx -unc Ssl -a install --domain $domain --chain 0
+  fi
+
   echo -e "[${BRIGHT}${GREEN}RELOAD${NORMAL}] .. SSL update successful\n"
   echo -e "\nhttps://www.sslshopper.com/ssl-checker.html#hostname=${domain}\n"
-
 fi
 }
